+++ src/Cooked/MockChain/Write.hs
@@ -214,11 +214,11 @@ runMockChainWrite = interpret $ \case
         -- We retrieve the utxos created by the transaction
         let utxos = P.Ledger.fromCardanoTxIn . snd <$> P.Ledger.getCardanoTxOutRefs cardanoTx
         -- We combine them with their corresponding `TxSkelOut`
-        let newOutputs = zip utxos (txSkelOuts finalTxSkel)
+        let newOutputs = zip utxos (txSkelOutputs finalTxSkel)
         -- We add the news utxos to the state
         forM_ newOutputs $ modify' . uncurry addOutput
         -- And remove the old ones
-        forM_ (Map.toList $ txSkelIns finalTxSkel) $ modify' . removeOutput . fst
+        forM_ (Map.toList $ txSkelInputs finalTxSkel) $ modify' . removeOutput . fst
         -- We return the newly created outputs
         return newOutputs
       -- This is a theoretical unreachable case. Since we fail in Phase 2, it